Refactor schemaTypes.ts into more specific places
