+++ packages/dds/tree/src/simple-tree/api/conciseTree.ts
@@ -8,7 +8,7 @@ import type { IFluidHandle } from "@fluidframework/core-interfaces";
 import type { ITreeCursor } from "../../core/index.js";
 import type { TreeNodeSchema } from "../core/index.js";
 import { getUnhydratedContext } from "../createContext.js";
-import type { TreeLeafValue, ImplicitAllowedTypes } from "../schemaTypes.js";
+import type { TreeLeafValue, ImplicitAllowedTypes } from "../fieldSchema.js";
 
 import {
 	customFromCursor,
