+++ packages/dds/tree/src/simple-tree/api/identifierIndex.ts
@@ -5,7 +5,7 @@
 
 import { UsageError } from "@fluidframework/telemetry-utils/internal";
 import type { TreeIndexKey } from "../../feature-libraries/index.js";
-import { FieldKind, type ImplicitFieldSchema } from "../schemaTypes.js";
+import { FieldKind, type ImplicitFieldSchema } from "../fieldSchema.js";
 import type { TreeNode } from "../core/index.js";
 import { ObjectNodeSchema } from "../node-kinds/index.js";
 import type { TreeView } from "./tree.js";
