+++ packages/dds/tree/src/simple-tree/api/schemaCompatibilityTester.ts
@@ -23,7 +23,7 @@ import {
 	fieldRealizer,
 	comparePosetElements,
 } from "../../feature-libraries/index.js";
-import type { FieldSchema } from "../schemaTypes.js";
+import type { FieldSchema } from "../fieldSchema.js";
 import { toStoredSchema } from "../toStoredSchema.js";
 
 import type { SchemaCompatibilityStatus } from "./tree.js";
