+++ packages/dds/tree/src/simple-tree/api/schemaCreationUtilities.ts
@@ -7,7 +7,7 @@ import { UsageError } from "@fluidframework/telemetry-utils/internal";
 import { assert, fail } from "@fluidframework/core-utils/internal";
 
 import type { SchemaFactory, ScopedSchemaName } from "./schemaFactory.js";
-import type { NodeFromSchema } from "../schemaTypes.js";
+import type { NodeFromSchema } from "../fieldSchema.js";
 import type {
 	InternalTreeNode,
 	NodeKind,
