+++ packages/dds/tree/src/simple-tree/core/context.ts
@@ -9,8 +9,9 @@ import type {
 	FlexTreeHydratedContext,
 } from "../../feature-libraries/index.js";
 import { brand } from "../../util/index.js";
+import type { NormalizedAnnotatedAllowedTypes } from "./allowedTypes.js";
 
-import type { NormalizedAnnotatedAllowedTypes, TreeNodeSchema } from "./treeNodeSchema.js";
+import type { TreeNodeSchema } from "./treeNodeSchema.js";
 import { walkAllowedTypes } from "./walkSchema.js";
 
 /**
