+++ packages/dds/tree/src/simple-tree/core/getOrCreateNode.ts
@@ -5,7 +5,7 @@
 
 import type { TreeValue } from "../../core/index.js";
 import { isFlexTreeNode, type FlexTreeUnknownUnboxed } from "../../feature-libraries/index.js";
-import type { TreeLeafValue } from "../schemaTypes.js";
+import type { TreeLeafValue } from "../fieldSchema.js";
 
 import type { TreeNode } from "./treeNode.js";
 import {
