+++ packages/dds/tree/src/simple-tree/core/index.ts
@@ -24,17 +24,34 @@ export {
 	privateToken,
 	inPrototypeChain,
 } from "./treeNode.js";
-export {
-	type TreeNodeSchema,
-	NodeKind,
-	type TreeNodeSchemaClass,
-	type TreeNodeSchemaNonClass,
-	type TreeNodeSchemaCore,
-	type TreeNodeSchemaBoth,
-	type AnnotatedAllowedType,
-	type NormalizedAnnotatedAllowedTypes,
-	isAnnotatedAllowedTypes,
+export { NodeKind } from "./treeNodeSchema.js";
+export type {
+	TreeNodeSchema,
+	TreeNodeSchemaClass,
+	TreeNodeSchemaNonClass,
+	TreeNodeSchemaCore,
+	TreeNodeSchemaBoth,
+	NodeSchemaMetadata,
+	TreeLeafValue,
 } from "./treeNodeSchema.js";
+export {
+	isAnnotatedAllowedTypes,
+	isAnnotatedAllowedType,
+	normalizeAllowedTypes,
+	normalizeAnnotatedAllowedTypes,
+	unannotateImplicitAllowedTypes,
+} from "./allowedTypes.js";
+export type {
+	AnnotatedAllowedType,
+	NormalizedAnnotatedAllowedTypes,
+	ImplicitAllowedTypes,
+	ImplicitAnnotatedAllowedTypes,
+	UnannotateImplicitAllowedTypes,
+	AllowedTypesMetadata,
+	AllowedTypes,
+	TreeNodeFromImplicitAllowedTypes,
+	InsertableTreeNodeFromImplicitAllowedTypes,
+} from "./allowedTypes.js";
 export { walkAllowedTypes, type SchemaVisitor } from "./walkSchema.js";
 export { Context, HydratedContext, SimpleContextSlot } from "./context.js";
 export {
