+++ packages/dds/tree/src/simple-tree/core/unhydratedFlexTree.ts
@@ -54,7 +54,7 @@ import {
 import { brand, filterIterable, getOrCreate, mapIterable } from "../../util/index.js";
 
 import type { Context } from "./context.js";
-import type { ContextualFieldProvider } from "../schemaTypes.js";
+import type { ContextualFieldProvider } from "../fieldSchema.js";
 import type { TreeNode } from "./treeNode.js";
 
 interface UnhydratedTreeSequenceFieldEditBuilder
