+++ packages/dds/tree/src/simple-tree/node-kinds/object/objectNodeTypes.ts
@@ -12,8 +12,8 @@ import type {
 import type {
 	FieldSchemaAlpha,
 	ImplicitAnnotatedFieldSchema,
-	UnannotateSchemaRecord,
-} from "../../schemaTypes.js";
+	UnannotateImplicitFieldSchema,
+} from "../../fieldSchema.js";
 import { NodeKind, type TreeNodeSchemaClass, type TreeNodeSchema } from "../../core/index.js";
 import type { FieldKey } from "../../../core/index.js";
 import type { SimpleObjectFieldSchema, SimpleObjectNodeSchema } from "../../simpleSchema.js";
@@ -45,6 +45,16 @@ export interface ObjectNodeSchema<
 	 */
 	readonly fields: ReadonlyMap<string, FieldSchemaAlpha & SimpleObjectFieldSchema>;
 }
+
+/**
+ * Removes annotations from field schemas in a schema record.
+ * @system @alpha
+ */
+export type UnannotateSchemaRecord<
+	T extends RestrictiveStringRecord<ImplicitAnnotatedFieldSchema>,
+> = {
+	readonly [P in Extract<keyof T, string>]: UnannotateImplicitFieldSchema<T[P]>;
+};
 
 /**
  * Extra data provided on all {@link ObjectNodeSchema} that is not included in the (soon possibly public) ObjectNodeSchema type.
