+++ packages/dds/tree/src/simple-tree/node-kinds/record/recordNode.ts
@@ -20,21 +20,18 @@ import {
 	getKernel,
 	type InternalTreeNode,
 	type NormalizedAnnotatedAllowedTypes,
+	type NodeSchemaMetadata,
+	type ImplicitAnnotatedAllowedTypes,
+	type UnannotateImplicitAllowedTypes,
+	type ImplicitAllowedTypes,
+	normalizeAllowedTypes,
+	normalizeAnnotatedAllowedTypes,
+	unannotateImplicitAllowedTypes,
+	type TreeNodeFromImplicitAllowedTypes,
 } from "../../core/index.js";
 import { getUnhydratedContext } from "../../createContext.js";
 import { tryGetTreeNodeForField } from "../../getTreeNodeForField.js";
-import {
-	type NodeSchemaMetadata,
-	type TreeNodeFromImplicitAllowedTypes,
-	type ImplicitAllowedTypes,
-	normalizeAllowedTypes,
-	unannotateImplicitAllowedTypes,
-	type ImplicitAnnotatedAllowedTypes,
-	type UnannotateImplicitAllowedTypes,
-	createFieldSchema,
-	FieldKind,
-	normalizeAnnotatedAllowedTypes,
-} from "../../schemaTypes.js";
+import { createFieldSchema, FieldKind } from "../../fieldSchema.js";
 import {
 	unhydratedFlexTreeFromInsertable,
 	type InsertableContent,
