+++ packages/dds/tree/src/simple-tree/walkFieldSchema.ts
@@ -4,7 +4,7 @@
  */
 
 import { type TreeNodeSchema, walkAllowedTypes, type SchemaVisitor } from "./core/index.js";
-import { type ImplicitFieldSchema, normalizeFieldSchema } from "./schemaTypes.js";
+import { type ImplicitFieldSchema, normalizeFieldSchema } from "./fieldSchema.js";
 
 /**
  * Traverses all {@link TreeNodeSchema} schema reachable from `schema`, applying the visitor pattern.
