+++ packages/dds/tree/src/test/shared-tree/fuzz/fuzzUtils.ts
@@ -34,7 +34,7 @@ import { expectEqualPaths, SharedTreeTestFactory } from "../../utils.js";
 import type {
 	NodeBuilderData,
 	// eslint-disable-next-line import/no-internal-modules
-} from "../../../simple-tree/schemaTypes.js";
+} from "../../../simple-tree/fieldSchema.js";
 import {
 	SchemaFactory,
 	TreeViewConfiguration,
