+++ packages/dds/tree/src/test/simple-tree/api/schemaFactory.spec.ts
@@ -43,7 +43,7 @@ import type {
 	TreeFieldFromImplicitField,
 	TreeNodeFromImplicitAllowedTypes,
 	// eslint-disable-next-line import/no-internal-modules
-} from "../../../simple-tree/schemaTypes.js";
+} from "../../../simple-tree/fieldSchema.js";
 import type {
 	areSafelyAssignable,
 	requireAssignableTo,
