+++ packages/dds/tree/src/test/simple-tree/schemaTypes.spec.ts
@@ -54,7 +54,7 @@ import {
 	unannotateImplicitAllowedTypes,
 	unannotateSchemaRecord,
 	// eslint-disable-next-line import/no-internal-modules
-} from "../../simple-tree/schemaTypes.js";
+} from "../../simple-tree/fieldSchema.js";
 import type {
 	areSafelyAssignable,
 	isAssignableTo,
