+++ packages/dds/tree/src/test/simple-tree/unhydratedNode.spec.ts
@@ -22,7 +22,7 @@ import type {
 	ContextualFieldProvider,
 	FieldProvider,
 	// eslint-disable-next-line import/no-internal-modules
-} from "../../simple-tree/schemaTypes.js";
+} from "../../simple-tree/fieldSchema.js";
 import { hydrate } from "./utils.js";
 import { TreeStatus } from "../../feature-libraries/index.js";
 import { validateUsageError } from "../utils.js";
